Do not import process-message-content in frontend
