+++ api/mutations/message/editMessage.js
@@ -19,7 +19,7 @@ import { validateRawContentState } from '../../utils/validate-draft-js-input';
 import processMessageContent, {
   messageTypeObj,
 } from 'shared/draft-utils/process-message-content';
-import type { MessageType } from 'shared/draft-utils/process-message-content';
+import type { MessageType } from 'shared/draft-utils/message-types';
 
 type Args = {
   input: {
