+++ shared/draft-utils/message-types.js
@@ -1,0 +1,8 @@
+// @flow
+
+export const messageTypeObj = {
+  text: 'text',
+  media: 'media',
+  draftjs: 'draftjs',
+};
+export type MessageType = $Keys<typeof messageTypeObj>;
